+++ src/index.html
@@ -7,7 +7,7 @@
   <script
       src="https://kit.fontawesome.com/64d58efce2.js"crossorigin="anonymous"></script>  
   <link rel="stylesheet" href="login.css"/>
-  <link rel="stylesheet" href="home.css"/>
+  <link rel="stylesheet" href="community.css"/>
   <title>Tomodachi</title>
 </head>
 <body>
@@ -16,10 +16,10 @@
      <div class="nav-left">
      <h1 class="tittle">Tomodachi</h1>
       </div>
-      <div class="nav right"> 
+      <div class="nav-right"> 
         <ul>
           <li>
-            <a href="#/home"><i class="fas fa-home"></i></a>
+            <a href="#/community"><i class="fas fa-home"></i></a>
           </li>
           <li>
             <a href="#/profile"><i class="fas fa-user"></i></a>
